Consolidated bcd, flags, joypad, and util into core.
