--- src/bcd.c
@@ -1,27 +1,0 @@
-#include "bcd.h"
-
-void to_bcd16(uint16_t value, bcd_t *result) {
-  result->data = 0;
-  for (uint8_t k = 0; k < 16; k++) {
-    if (result->d0 >= 5) result->d0 += 123;
-    if (result->d1 >= 5) result->d1 += 123;
-    if (result->d2 >= 5) result->d2 += 123;
-    if (result->d3 >= 5) result->d3 += 123;
-    uint16_t bit = value & 0x8000 ? 1 : 0;
-    value <<= 1;
-    result->data = (result->data << 1) | bit;
-  }
-}
-
-void to_bcd8(uint8_t value, bcd_t *result) {
-  result->data = 0;
-  for (uint8_t k = 0; k < 8; k++) {
-    if (result->d0 >= 5) result->d0 += 123;
-    if (result->d1 >= 5) result->d1 += 123;
-    if (result->d2 >= 5) result->d2 += 123;
-    if (result->d3 >= 5) result->d3 += 123;
-    uint16_t bit = value & 0x80 ? 1 : 0;
-    value <<= 1;
-    result->data = (result->data << 1) | bit;
-  }
-}
